More misc size_t issues

+++ rpmio/rpmio.c
@@ -817,7 +817,7 @@ static struct FDIO_s ufdio_s = {
 };
 FDIO_t ufdio = &ufdio_s ;
 
-int timedRead(FD_t fd, void * bufptr, int length)
+ssize_t timedRead(FD_t fd, void * bufptr, size_t length)
 {
     return ufdio->read(fd, bufptr, length);
 }
